Major changes
--------------------------
==Information=============
--------------------------

Major changes to the whole unity project, this push will casue problems
with other updates to the project, contributers needs to check if their
code is missing and add this again.

--------------------------
==What's new==============
--------------------------
Removed all scripts.

//Created:

*BrickCont - This script is located on all of the coloured bricks.
This controls the downwards-movement of each brick, and
has collision-detection added.

*Controller - New controller scipt that is added to the new
GameObject called "SpawnBrick". Reason for this is simply
because we needed a dedicated controller script.

*Instantiater - This script has also been added to the
"SpawnBrick"-GameObject.
This script will handle the creation of the different coloured
bricks.

*Spawn  - This script is very similiar to the script located on previous
versions on the ButtomLayer-prefab. This script will spawn the
gameObject
"SpawnBrick".

This has been done because we needed a better solution to spawn random
bricks within a GameObject.

Commented most of the code in every script!

--------------------------
==Bugs====================
--------------------------
Spawn script doesn't see the Active Block resulting in a massive
clone-war, this is VERY URGENT and need to be fixxed.